fix(aleo_ffi)!: guard node-driven import recursion; widen amount/fee ABI to u64
Review round 8, fixed as classes rather than instances:

Untrusted-node class:
- add_program_from_node now threads the in-progress import chain
  through the recursion: an import cycle reported by the node
  (a -> b -> a, impossible on-chain, so only a lying node) is rejected
  instead of recursed into until the stack overflows; chain depth is
  capped at MAX_IMPORT_DEPTH (16); and the fetched source must declare
  the requested program ID, so a node cannot substitute a different
  program. Both guards fire before any HTTP, which makes them
  offline-testable (2 new tests).

Integer-across-the-ABI class (BREAKING for the C ABI):
- Every amount/fee parameter goes from c_int to u64, and get_base_fee
  returns u64: snarkVM fees/amounts are u64 microcredits, so i32 could
  not express legal values past ~2147 credits and a negative or
  truncated value became an enormous fee via `as u64`. Dart typedefs
  move from ffi.Int to ffi.Uint64 in step (both sides ship together
  from this repo), and the Dart API rejects negative amounts/fees
  (AleoUtils.checkAmount) since a negative Dart int would wrap.
  transfer_inputs takes u64; the `as i64`/`as u64` casts are gone.
- Same class, found by sweep: sign_message/verify rejected a negative
  message length instead of wrapping it to a huge usize and reading
  out of bounds (UB); seed_to_private_key/sign_message/verify also
  null-check their byte pointers.

Tests: amount_units_test now drives 2^31, 2^32 and 10^15 microcredit
amounts through the real cdylib in CI (would have caught the i32
truncation) and asserts negative inputs throw before the FFI; Rust
adds negative-length, import-cycle and import-depth tests (11 offline
total).

Verified locally: cargo test --lib 11 passed; dart test against the
rebuilt dylib 20 passed (2 manual live-node tests skipped); dart
analyze clean.

+++ packages/aleo_dart/lib/src/rust_lib/programs_rust_ffi.dart
@@ -4,12 +4,14 @@ import 'package:ffi/ffi.dart';
 typedef TypeTestInRust = ffi.Int Function(ffi.Int, ffi.Int);
 typedef TypeTestInDart = int Function(int, int);
 
+// Amounts and fees cross the ABI as unsigned 64-bit microcredits (snarkVM's
+// fee/amount type); 32-bit types cannot express the legal range.
 typedef TypeTransferInRust = ffi.Pointer<Utf8> Function(
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
-    ffi.Int,
-    ffi.Int,
+    ffi.Uint64,
+    ffi.Uint64,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
@@ -33,7 +35,7 @@ typedef TypeAuthorizationInRust = ffi.Pointer<Utf8> Function(
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
-    ffi.Int,
+    ffi.Uint64,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>);
@@ -66,7 +68,7 @@ typedef TypeJoinInRust = ffi.Pointer<Utf8> Function(
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
-    ffi.Int,
+    ffi.Uint64,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>);
@@ -83,7 +85,7 @@ typedef TypeJoinInDart = ffi.Pointer<Utf8> Function(
 typedef TypeJoinAuthorization = ffi.Pointer<Utf8> Function(ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>, ffi.Pointer<Utf8>, ffi.Pointer<Utf8>, ffi.Pointer<Utf8>);
 
-typedef TypeGetBaseFeeInRust = ffi.Int Function(
+typedef TypeGetBaseFeeInRust = ffi.Uint64 Function(
   ffi.Pointer<Utf8>,
   ffi.Pointer<Utf8>,
   ffi.Pointer<Utf8>,
@@ -116,7 +118,7 @@ typedef TypeExecuteProgramInRust = ffi.Pointer<Utf8> Function(
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
-    ffi.Int,
+    ffi.Uint64,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>);
 
@@ -131,7 +133,7 @@ typedef TypeExecuteProgramInDart = ffi.Pointer<Utf8> Function(
 
 typedef TypeContractFeeExecutionInRust = ffi.Pointer<Utf8> Function(
     ffi.Pointer<Utf8>,
-    ffi.Int,
+    ffi.Uint64,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
     ffi.Pointer<Utf8>,
